Refactor - Changed folder structure
