+++ src/main/java/com/sentinels/robot/Main.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot;
+package com.sentinels.robot;
 
 import edu.wpi.first.wpilibj.RobotBase;
