+++ src/main/java/com/sentinels/robot/commands/armmech/arm/ArmExtend.java
@@ -2,13 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.armmech.arm;
+package com.sentinels.robot.commands.armmech.arm;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.arm.Arm;
+import com.sentinels.robot.subsystems.arm.Arm;
 
 /** An example command that uses an example subsystem. */
-public class ArmPivot extends CommandBase {
+public class ArmExtend extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
   private final Arm m_ArmSubsystem;
 
@@ -17,7 +17,7 @@ public class ArmPivot extends CommandBase {
    *
    * @param subsystem The subsystem used by this command.
    */
-  public ArmPivot(Arm subsystem) {
+  public ArmExtend(Arm subsystem) {
     m_ArmSubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
