+++ src/main/java/com/sentinels/robot/commands/armmech/arm/ArmRetract.java
@@ -2,47 +2,34 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.drivetrain;
+package com.sentinels.robot.commands.armmech.arm;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import frc.robot.subsystems.drive.Drivetrain;
+import com.sentinels.robot.subsystems.arm.Arm;
 
 /** An example command that uses an example subsystem. */
-public class DrivetrainDrive extends CommandBase {
+public class ArmRetract extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private Drivetrain m_DrivetrainSubsystem;
-  private CommandXboxController controller;
+  private final Arm m_ArmSubsystem;
 
-  double left;
-  double right;
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public DrivetrainDrive(Drivetrain subsystem, CommandXboxController controller) {
-    m_DrivetrainSubsystem = subsystem;
-    this.controller = controller;
+  public ArmRetract(Arm subsystem) {
+    m_ArmSubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {
-    left = 0;
-    right = 0;
-  }
+  public void initialize() {}
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {
-    left = controller.getLeftY();
-    right = controller.getRightY();
-
-    m_DrivetrainSubsystem.setSpeed(left, right);
-  }
+  public void execute() {}
 
   // Called once the command ends or is interrupted.
   @Override
