+++ src/main/java/com/sentinels/robot/commands/armmech/intake/IntakeOpen.java
@@ -2,23 +2,23 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.armmech.arm;
+package com.sentinels.robot.commands.armmech.intake;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.arm.Arm;
+import com.sentinels.robot.subsystems.arm.ArmIntake;
 
 /** An example command that uses an example subsystem. */
-public class ArmExtend extends CommandBase {
+public class IntakeOpen extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final Arm m_ArmSubsystem;
+  private final ArmIntake m_ArmIntakeSubsystem;
 
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public ArmExtend(Arm subsystem) {
-    m_ArmSubsystem = subsystem;
+  public IntakeOpen(ArmIntake subsystem) {
+    m_ArmIntakeSubsystem = subsystem;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
   }
