+++ src/main/java/com/sentinels/robot/commands/autonomous/Autos.java
@@ -2,13 +2,13 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.autonomous;
+package com.sentinels.robot.commands.autonomous;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.Commands;
 
 // Import all subsystems for auton use
-import frc.robot.subsystems.*;
+import com.sentinels.robot.subsystems.*;
 
 public final class Autos {
