+++ src/main/java/com/sentinels/robot/commands/drivetrain/DrivetrainDrive.java
@@ -2,34 +2,47 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.armmech.intake;
+package com.sentinels.robot.commands.drivetrain;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.arm.ArmIntake;
+import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
+import com.sentinels.robot.subsystems.drive.Drivetrain;
 
 /** An example command that uses an example subsystem. */
-public class IntakeOpen extends CommandBase {
+public class DrivetrainDrive extends CommandBase {
   @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final ArmIntake m_ArmIntakeSubsystem;
+  private Drivetrain m_DrivetrainSubsystem;
+  private CommandXboxController controller;
 
+  double left;
+  double right;
   /**
    * Creates a new ExampleCommand.
    *
    * @param subsystem The subsystem used by this command.
    */
-  public IntakeOpen(ArmIntake subsystem) {
-    m_ArmIntakeSubsystem = subsystem;
+  public DrivetrainDrive(Drivetrain subsystem, CommandXboxController controller) {
+    m_DrivetrainSubsystem = subsystem;
+    this.controller = controller;
     // Use addRequirements() here to declare subsystem dependencies.
     addRequirements(subsystem);
   }
 
   // Called when the command is initially scheduled.
   @Override
-  public void initialize() {}
+  public void initialize() {
+    left = 0;
+    right = 0;
+  }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
-  public void execute() {}
+  public void execute() {
+    left = controller.getLeftY();
+    right = controller.getRightY();
+
+    m_DrivetrainSubsystem.setSpeed(left, right);
+  }
 
   // Called once the command ends or is interrupted.
   @Override
