+++ src/main/java/com/sentinels/robot/commands/drivetrain/DrivetrainStop.java
@@ -2,10 +2,10 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.drivetrain;
+package com.sentinels.robot.commands.drivetrain;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.drive.Drivetrain;
+import com.sentinels.robot.subsystems.drive.Drivetrain;
 
 /** An example command that uses an example subsystem. */
 public class DrivetrainStop extends CommandBase {
