+++ src/main/java/com/sentinels/robot/constants/Motors.java
@@ -6,7 +6,7 @@
     Copyright (c) 2023 The Sentinels. All rights reserved.
 ***************************************************************/
 
-package frc.robot.constants;
+package com.sentinels.robot.constants;
 
 // Configurations for all motors will be defined in this file.
